makes separate room components
