+++ src/constants/socket-context.ts
@@ -1,5 +1,5 @@
 import openSocket from 'socket.io-client';
-import Constants from '../constants/Constants';
+import Constants from './Constants';
 
 const socket = openSocket(`${Constants.SERVER_HOST}:${Constants.SERVER_PORT}`);
